requirements.txt
imapclient==2.3.1
email-validator==2.1.1
python-dotenv==1.0.1
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
numpy==1.26.4
transformers==4.37.2
accelerate==0.26.1